Moved logic constans to Config
